+++ src/Pages/RecruitPage/data.ts
@@ -1,0 +1,17 @@
+export const recruitData = {
+    form: '/25/2025학년도 PRIMITIVE 신청서.hwp',
+    ot: '/25/2025 PRIMITIVE 신입생 OT 자료.pdf',
+    recruitNotice: '/25/25PrimitiveRecruit.webp',
+    presidentData: {
+        position: '회장',
+        name: '장준우',
+        kakaoID: 'JJW1102',
+        email: 'jwjang1102@naver.com'
+    },
+    vicePresidentData: {
+        position: '부회장',
+        name: '최세진',
+        kakaoID: 'marin6670',
+        email: 'marin6670@gmail.com'
+    }
+}
