+++ vite.config.js
@@ -1,6 +1,8 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import tailwindcss from "tailwindcss";
+import path from "path";
+
 export default defineConfig({
   plugins: [react(), tailwindcss()],
   //build 시에 모든 console.log를 제거
@@ -13,4 +15,9 @@ export default defineConfig({
       },
     },
   },
+  resolve: {
+    alias: {
+      '@': path.join(__dirname, '/')
+    },
+  }
 });
